Move query parameter parsing to QueryParameter
