+++ flow-server/src/test/java/com/vaadin/flow/router/QueryParametersTest.java
@@ -32,9 +32,9 @@ import org.junit.Test;
 
 public class QueryParametersTest {
 
-    private String simpleInputQueryString = "one=1&two=2&three=3&four&five=4%2f5";
+    private String simpleInputQueryString = "one=1&two=2&three=3&four&five=4%2f5%266";
 
-    private String complexInputQueryString = "one=1&one=11&two=2&two=22&three=3&four&five=4%2f5";
+    private String complexInputQueryString = "one=1&one=11&two=2&two=22&three=3&four&five=4%2f5%266";
 
     private Map<String, String> getSimpleInputParameters() {
         Map<String, String> inputParameters = new HashMap<>();
@@ -105,7 +105,7 @@ public class QueryParametersTest {
         expectedFullParams.put("two", Collections.singletonList("2"));
         expectedFullParams.put("three", Collections.singletonList("3"));
         expectedFullParams.put("four", Collections.singletonList(""));
-        expectedFullParams.put("five", Collections.singletonList("4/5"));
+        expectedFullParams.put("five", Collections.singletonList("4/5&6"));
         assertEquals(expectedFullParams, simpleParams.getParameters());
     }
 
@@ -157,7 +157,7 @@ public class QueryParametersTest {
         expectedFullParams.put("two", Arrays.asList("2", "22"));
         expectedFullParams.put("three", Collections.singletonList("3"));
         expectedFullParams.put("four", Collections.singletonList(""));
-        expectedFullParams.put("five", Collections.singletonList("4/5"));
+        expectedFullParams.put("five", Collections.singletonList("4/5&6"));
         assertEquals(expectedFullParams, fullParams.getParameters());
     }
